+++ src/containers/Application.js
@@ -508,7 +508,12 @@ class Application extends Component {
 						<div className="app-body">
 							{
 								(this.props.hasSidebar || typeof this.props.hasSidebar === 'undefined') &&
-								<Sidebar app={this} navigation={this.Navigation} display="lg" />
+									<Sidebar
+										app={this}
+										navigation={this.Navigation}
+										display="lg"
+										sidebarItemsOrder={this.props.sidebarItemsOrder}
+									/>
 							}
 							<Main hasSidebar={this.props.hasSidebar}>
 								<Suspense